moved communicate_with_progress from utils.py to nmap3.py so it has reference to self.xml_path. implemented io from xml_path file
