+++ nmap3/utils.py
@@ -24,8 +24,6 @@ import sys
 import os
 import ctypes
 import functools
-import threading
-import queue
 import re
 from typing import Optional, Callable
 
@@ -105,82 +103,8 @@ def nmap_is_installed_async():
         return wrapped
     return  wrapper 
 
-def communicate_with_progress(
-    sub_proc: subprocess.Popen,
-    timeout: int = None,
-    progress_callback: Optional[Callable[[float], None]] = None
-) -> tuple[str, str]:
-    """
-    Reads stdout and stderr from a subprocess, optionally reporting progress.
-
-    Parameters
-    ----------
-    sub_proc : subprocess.Popen
-        The subprocess object to communicate with.
-    timeout : int | None, optional
-        Timeout in seconds for the subprocess. If None, waits until finished.
-    progress_callback : Callable[[float], None] | None, optional
-        A callback function that is called with the current scan progress
-        as a float between 0.0 and 100.0. Called whenever a line
-        matching "<number>% done" is read from stdout.
-
-    Returns
-    -------
-    Tuple[str, str]
-        A tuple containing the full stdout output and stderr output.
-
-    Example
-    -------
-    def my_progress(progress: float):
-        print(f"Progress: {progress:.2f}%")
-
-    import subprocess
-
-    proc = subprocess.Popen(["/usr/bin/nmap", "-oX", "-", "example.com"],
-                            stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE,
-                            text=True)
-    
-    output, errs = communicate_with_progress(proc, timeout=60, progress_callback=my_progress)
-    """
-    stdout_queue = queue.Queue()
-    stderr_queue = queue.Queue()
-
-    def reader(pipe, q):
-        for line in pipe:
-            q.put(line)
-        pipe.close()
-
-    #start threads to read stdout and stderr
-    t_out = threading.Thread(target=reader, args=(sub_proc.stdout, stdout_queue))
-    t_err = threading.Thread(target=reader, args=(sub_proc.stderr, stderr_queue))
-    t_out.start()
-    t_err.start()
-
+def read_xml_file(path:str) -> str:
     output = ""
-    errs = ""
-
-    while True:
-    
-        if sub_proc.poll() is not None and stdout_queue.empty() and stderr_queue.empty():
-            break #once sub_proc terminates and stdout_queue and stderr_queue are empty we are done
-
-        #Process stdout
-        while not stdout_queue.empty():
-            line = stdout_queue.get_nowait()
-            if progress_callback:
-                #grab the progress from stdout and pass to progress_callback
-                match = re.search(r'(\d+(?:\.\d+)?)% done', line)
-                if match:
-                    progress = float(match.group(1))
-                    progress_callback(progress)
-
-        #Process stderr
-        while not stderr_queue.empty():
-            line = stderr_queue.get_nowait()
-            errs += line
-
-    with open("tmp") as f: #TODO: replace
+    with open(path) as f:
         output = f.read()
-
-    return output, errs
+    return output
